Added 5 folders and moved files origanlly in the disease folder to the rest of the folders
