+++ src/main/java/com/example/demo/DemoApplication.java
@@ -1,16 +1,8 @@
 package com.example.demo;
 
-import com.example.demo.disease.Cancers;
-import com.example.demo.disease.Cancerslist;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
-
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.Month;
-import java.util.List;
 
 @SpringBootApplication
 @RestController
