+++ src/main/java/com/example/demo/config/CancersConfig.java
@@ -1,5 +1,7 @@
-package com.example.demo.disease;
+package com.example.demo.config;
 
+import com.example.demo.io.Cancers;
+import com.example.demo.repository.CancersRepository;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
