+++ src/main/java/com/example/demo/disease/Cancerslist.java
@@ -2,6 +2,8 @@ package com.example.demo.disease;
 
 
 
+import com.example.demo.io.Cancers;
+
 import java.util.List;
 import java.util.Map;
