+++ src/main/java/com/example/demo/io/Cancers.java
@@ -1,4 +1,4 @@
-package com.example.demo.disease;
+package com.example.demo.io;
 
 import jakarta.persistence.*;
