+++ src/main/java/com/example/demo/repository/CancersRepository.java
@@ -1,5 +1,6 @@
-package com.example.demo.disease;
+package com.example.demo.repository;
 
+import com.example.demo.io.Cancers;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
